Replace UTXO standard impl from structs to utxo_common mod
