+++ mm2src/coins/qtum.rs
@@ -7,9 +7,9 @@ use common::jsonrpc_client::{JsonRpcClient, JsonRpcRequest, JsonRpcError, RpcRes
 use crate::{HistorySyncState, FoundSwapTxSpend, MarketCoinOps, MmCoin, SwapOps, TradeFee, TradeInfo,
             TransactionDetails, TransactionEnum, TransactionFut, WithdrawRequest};
 use crate::eth::{ERC20_CONTRACT, u256_to_big_decimal, wei_from_big_decimal};
-use crate::utxo::{ActualTxFee, AdditionalTxData, FeePolicy, UtxoArc, UtxoArcCommonOps, UtxoArcCommonOpsStandardImpl,
-                  utxo_arc_from_conf_and_request, UtxoCoinCommonOps, UtxoCoinCommonOpsStandardImpl,
-                  UtxoMarketCoinOpsStandardImpl, UtxoMmCoinOpsStandardImpl, UtxoSwapOpsStandardImpl, withdraw_impl};
+use crate::utxo::{ActualTxFee, AdditionalTxData, FeePolicy, UtxoArc, UtxoArcCommonOps,
+                  utxo_arc_from_conf_and_request, UtxoCoinCommonOps};
+use crate::utxo::utxo_common;
 use crate::utxo::rpc_clients::{ElectrumClient, UnspentInfo, UtxoRpcClientEnum};
 use ethabi::Token;
 use ethereum_types::{H160, U256};
@@ -90,19 +90,19 @@ impl Deref for Qrc20Coin {
 #[async_trait]
 impl UtxoCoinCommonOps for Qrc20Coin {
     async fn get_tx_fee(&self) -> Result<ActualTxFee, JsonRpcError> {
-        UtxoCoinCommonOpsStandardImpl::get_tx_fee(&self).await
+        utxo_common::get_tx_fee(&self).await
     }
 
     async fn get_htlc_spend_fee(&self) -> Result<u64, String> {
-        UtxoCoinCommonOpsStandardImpl::get_htlc_spend_fee(self).await
+        utxo_common::get_htlc_spend_fee(self).await
     }
 
     fn addresses_from_script(&self, script: &Script) -> Result<Vec<Address>, String> {
-        UtxoCoinCommonOpsStandardImpl::addresses_from_script(&self, script)
+        utxo_common::addresses_from_script(&self, script)
     }
 
     fn denominate_satoshis(&self, satoshi: i64) -> f64 {
-        UtxoCoinCommonOpsStandardImpl::denominate_satoshis(&self, satoshi)
+        utxo_common::denominate_satoshis(&self, satoshi)
     }
 
     fn search_for_swap_tx_spend(
@@ -114,7 +114,7 @@ impl UtxoCoinCommonOps for Qrc20Coin {
         tx: &[u8],
         search_from_block: u64)
         -> Result<Option<FoundSwapTxSpend>, String> {
-        UtxoCoinCommonOpsStandardImpl::search_for_swap_tx_spend(
+        utxo_common::search_for_swap_tx_spend(
             &self,
             time_lock,
             first_pub,
@@ -132,7 +132,7 @@ impl UtxoCoinCommonOps for Qrc20Coin {
 #[async_trait]
 impl UtxoArcCommonOps for Qrc20Coin {
     fn send_outputs_from_my_address(&self, outputs: Vec<TransactionOutput>) -> TransactionFut {
-        UtxoArcCommonOpsStandardImpl::send_outputs_from_my_address(self.clone(), outputs)
+        utxo_common::send_outputs_from_my_address(self.clone(), outputs)
     }
 
     fn validate_payment(
@@ -144,7 +144,7 @@ impl UtxoArcCommonOps for Qrc20Coin {
         priv_bn_hash: &[u8],
         amount: BigDecimal)
         -> Box<dyn Future<Item=(), Error=String> + Send> {
-        UtxoArcCommonOpsStandardImpl::validate_payment(
+        utxo_common::validate_payment(
             self.inner.clone(),
             payment_tx,
             time_lock,
@@ -155,7 +155,7 @@ impl UtxoArcCommonOps for Qrc20Coin {
     }
 
     async fn generate_transaction(&self, utxos: Vec<UnspentInfo>, outputs: Vec<TransactionOutput>, fee_policy: FeePolicy, fee: Option<ActualTxFee>) -> Result<(TransactionInputSigner, AdditionalTxData), String> {
-        UtxoArcCommonOpsStandardImpl::generate_transaction(
+        utxo_common::generate_transaction(
             self,
             utxos,
             outputs,
@@ -169,7 +169,7 @@ impl UtxoArcCommonOps for Qrc20Coin {
         data: AdditionalTxData,
         my_script_pub: Bytes)
         -> Result<(TransactionInputSigner, AdditionalTxData), String> {
-        UtxoArcCommonOpsStandardImpl::calc_interest_if_required(
+        utxo_common::calc_interest_if_required(
             &self,
             unsigned,
             data,
@@ -184,7 +184,7 @@ impl UtxoArcCommonOps for Qrc20Coin {
         script_data: Script,
         sequence: u32)
         -> Result<Transaction, String> {
-        UtxoArcCommonOpsStandardImpl::p2sh_spending_tx(
+        utxo_common::p2sh_spending_tx(
             &self,
             prev_transaction,
             redeem_script,
@@ -196,81 +196,81 @@ impl UtxoArcCommonOps for Qrc20Coin {
 
 impl SwapOps for Qrc20Coin {
     fn send_taker_fee(&self, fee_addr: &[u8], amount: BigDecimal) -> TransactionFut {
-        UtxoSwapOpsStandardImpl::send_taker_fee(self, fee_addr, amount)
+        utxo_common::send_taker_fee(self, fee_addr, amount)
     }
 
     fn send_maker_payment(&self, time_lock: u32, taker_pub: &[u8], secret_hash: &[u8], amount: BigDecimal) -> TransactionFut {
-        UtxoSwapOpsStandardImpl::send_maker_payment(self.clone(), time_lock, taker_pub, secret_hash, amount)
+        utxo_common::send_maker_payment(self.clone(), time_lock, taker_pub, secret_hash, amount)
     }
 
     fn send_taker_payment(&self, time_lock: u32, maker_pub: &[u8], secret_hash: &[u8], amount: BigDecimal) -> TransactionFut {
-        UtxoSwapOpsStandardImpl::send_taker_payment(self.clone(), time_lock, maker_pub, secret_hash, amount)
+        utxo_common::send_taker_payment(self.clone(), time_lock, maker_pub, secret_hash, amount)
     }
 
     fn send_maker_spends_taker_payment(
         &self, taker_payment_tx: &[u8], time_lock: u32, taker_pub: &[u8], secret: &[u8])
         -> TransactionFut {
-        UtxoSwapOpsStandardImpl::send_maker_spends_taker_payment(
+        utxo_common::send_maker_spends_taker_payment(
             self.clone(), taker_payment_tx, time_lock, taker_pub, secret)
     }
 
     fn send_taker_spends_maker_payment(
         &self, maker_payment_tx: &[u8], time_lock: u32, maker_pub: &[u8], secret: &[u8])
         -> TransactionFut {
-        UtxoSwapOpsStandardImpl::send_taker_spends_maker_payment(
+        utxo_common::send_taker_spends_maker_payment(
             self.clone(), maker_payment_tx, time_lock, maker_pub, secret)
     }
 
     fn send_taker_refunds_payment(
         &self, taker_payment_tx: &[u8], time_lock: u32, maker_pub: &[u8], secret_hash: &[u8])
         -> TransactionFut {
-        UtxoSwapOpsStandardImpl::send_taker_refunds_payment(
+        utxo_common::send_taker_refunds_payment(
             self.clone(), taker_payment_tx, time_lock, maker_pub, secret_hash)
     }
 
     fn send_maker_refunds_payment(
         &self, maker_payment_tx: &[u8], time_lock: u32, taker_pub: &[u8], secret_hash: &[u8])
         -> TransactionFut {
-        UtxoSwapOpsStandardImpl::send_maker_refunds_payment(
+        utxo_common::send_maker_refunds_payment(
             self.clone(), maker_payment_tx, time_lock, taker_pub, secret_hash)
     }
 
     fn validate_fee(&self, fee_tx: &TransactionEnum, fee_addr: &[u8], amount: &BigDecimal)
                     -> Box<dyn Future<Item=(), Error=String> + Send> {
-        UtxoSwapOpsStandardImpl::validate_fee(self.inner.clone(), fee_tx, fee_addr, amount)
+        utxo_common::validate_fee(self.inner.clone(), fee_tx, fee_addr, amount)
     }
 
     fn validate_maker_payment(
         &self, payment_tx: &[u8], time_lock: u32, maker_pub: &[u8], priv_bn_hash: &[u8], amount: BigDecimal)
         -> Box<dyn Future<Item=(), Error=String> + Send> {
-        UtxoSwapOpsStandardImpl::validate_maker_payment(self, payment_tx, time_lock, maker_pub, priv_bn_hash, amount)
+        utxo_common::validate_maker_payment(self, payment_tx, time_lock, maker_pub, priv_bn_hash, amount)
     }
 
     fn validate_taker_payment(
         &self, payment_tx: &[u8], time_lock: u32, taker_pub: &[u8], priv_bn_hash: &[u8], amount: BigDecimal)
         -> Box<dyn Future<Item=(), Error=String> + Send> {
-        UtxoSwapOpsStandardImpl::validate_taker_payment(
+        utxo_common::validate_taker_payment(
             self, payment_tx, time_lock, taker_pub, priv_bn_hash, amount)
     }
 
     fn check_if_my_payment_sent(
         &self, time_lock: u32, other_pub: &[u8], secret_hash: &[u8], search_from_block: u64)
         -> Box<dyn Future<Item=Option<TransactionEnum>, Error=String> + Send> {
-        UtxoSwapOpsStandardImpl::check_if_my_payment_sent(
+        utxo_common::check_if_my_payment_sent(
             self.inner.clone(), time_lock, other_pub, secret_hash, search_from_block)
     }
 
     fn search_for_swap_tx_spend_my(
         &self, time_lock: u32, other_pub: &[u8], secret_hash: &[u8], tx: &[u8], search_from_block: u64)
         -> Result<Option<FoundSwapTxSpend>, String> {
-        UtxoSwapOpsStandardImpl::search_for_swap_tx_spend_my(
+        utxo_common::search_for_swap_tx_spend_my(
             self, time_lock, other_pub, secret_hash, tx, search_from_block)
     }
 
     fn search_for_swap_tx_spend_other(
         &self, time_lock: u32, other_pub: &[u8], secret_hash: &[u8], tx: &[u8], search_from_block: u64)
         -> Result<Option<FoundSwapTxSpend>, String> {
-        UtxoSwapOpsStandardImpl::search_for_swap_tx_spend_other(
+        utxo_common::search_for_swap_tx_spend_other(
             self, time_lock, other_pub, secret_hash, tx, search_from_block)
     }
 }
@@ -281,7 +281,7 @@ impl MarketCoinOps for Qrc20Coin {
     }
 
     fn my_address(&self) -> Cow<str> {
-        UtxoMarketCoinOpsStandardImpl::my_address(&self)
+        utxo_common::my_address(&self)
     }
 
     fn my_balance(&self) -> Box<dyn Future<Item=BigDecimal, Error=String> + Send> {
@@ -307,47 +307,47 @@ impl MarketCoinOps for Qrc20Coin {
     }
 
     fn send_raw_tx(&self, tx: &str) -> Box<dyn Future<Item=String, Error=String> + Send> {
-        UtxoMarketCoinOpsStandardImpl::send_raw_tx(&self, tx)
+        utxo_common::send_raw_tx(&self, tx)
     }
 
     fn wait_for_confirmations(&self, tx: &[u8], confirmations: u64, requires_nota: bool, wait_until: u64, check_every: u64)
                               -> Box<dyn Future<Item=(), Error=String> + Send> {
-        UtxoMarketCoinOpsStandardImpl::wait_for_confirmations(&self, tx, confirmations, requires_nota, wait_until, check_every)
+        utxo_common::wait_for_confirmations(&self, tx, confirmations, requires_nota, wait_until, check_every)
     }
 
     fn wait_for_tx_spend(&self, transaction: &[u8], wait_until: u64, from_block: u64) -> TransactionFut {
-        UtxoMarketCoinOpsStandardImpl::wait_for_tx_spend(&self, transaction, wait_until, from_block)
+        utxo_common::wait_for_tx_spend(&self, transaction, wait_until, from_block)
     }
 
     fn tx_enum_from_bytes(&self, bytes: &[u8]) -> Result<TransactionEnum, String> {
-        UtxoMarketCoinOpsStandardImpl::tx_enum_from_bytes(bytes)
+        utxo_common::tx_enum_from_bytes(bytes)
     }
 
     fn current_block(&self) -> Box<dyn Future<Item=u64, Error=String> + Send> {
-        UtxoMarketCoinOpsStandardImpl::current_block(&self)
+        utxo_common::current_block(&self)
     }
 
     fn address_from_pubkey_str(&self, pubkey: &str) -> Result<String, String> {
-        UtxoMarketCoinOpsStandardImpl::address_from_pubkey_str(&self, pubkey)
+        utxo_common::address_from_pubkey_str(&self, pubkey)
     }
 
     fn display_priv_key(&self) -> String {
-        UtxoMarketCoinOpsStandardImpl::display_priv_key(&self)
+        utxo_common::display_priv_key(&self)
     }
 }
 
 impl MmCoin for Qrc20Coin {
     fn is_asset_chain(&self) -> bool {
-        UtxoMmCoinOpsStandardImpl::is_asset_chain(&self)
+        utxo_common::is_asset_chain(&self)
     }
 
     fn check_i_have_enough_to_trade(&self, amount: &MmNumber, balance: &MmNumber, trade_info: TradeInfo)
                                     -> Box<dyn Future<Item=(), Error=String> + Send> {
-        UtxoMmCoinOpsStandardImpl::check_i_have_enough_to_trade(self.clone(), amount, balance, trade_info)
+        utxo_common::check_i_have_enough_to_trade(self.clone(), amount, balance, trade_info)
     }
 
     fn can_i_spend_other_payment(&self) -> Box<dyn Future<Item=(), Error=String> + Send> {
-        UtxoMmCoinOpsStandardImpl::can_i_spend_other_payment()
+        utxo_common::can_i_spend_other_payment()
     }
 
     fn withdraw(&self, req: WithdrawRequest) -> Box<dyn Future<Item=TransactionDetails, Error=String> + Send> {
@@ -365,46 +365,46 @@ impl MmCoin for Qrc20Coin {
             let script_pubkey = try_s!(generate_token_transfer_script_pubkey(
                 to_addr, amount, 2_500_000, 40, &coin.contract_address));
 
-            withdraw_impl(coin, req, script_pubkey).await
+            utxo_common::withdraw_impl(coin, req, script_pubkey).await
         }
 
         Box::new(withdraw_(self.clone(), req).boxed().compat())
     }
 
     fn decimals(&self) -> u8 {
-        UtxoMmCoinOpsStandardImpl::decimals(&self)
+        utxo_common::decimals(&self)
     }
 
     fn process_history_loop(&self, ctx: MmArc) {
-        UtxoMmCoinOpsStandardImpl::process_history_loop(self, ctx)
+        utxo_common::process_history_loop(self, ctx)
     }
 
     fn tx_details_by_hash(&self, hash: &[u8]) -> Box<dyn Future<Item=TransactionDetails, Error=String> + Send> {
-        UtxoMmCoinOpsStandardImpl::tx_details_by_hash(self.clone(), hash)
+        utxo_common::tx_details_by_hash(self.clone(), hash)
     }
 
     fn history_sync_status(&self) -> HistorySyncState {
-        UtxoMmCoinOpsStandardImpl::history_sync_status(&self)
+        utxo_common::history_sync_status(&self)
     }
 
     fn get_trade_fee(&self) -> Box<dyn Future<Item=TradeFee, Error=String> + Send> {
-        UtxoMmCoinOpsStandardImpl::get_trade_fee(self.clone())
+        utxo_common::get_trade_fee(self.clone())
     }
 
     fn required_confirmations(&self) -> u64 {
-        UtxoMmCoinOpsStandardImpl::required_confirmations(&self)
+        utxo_common::required_confirmations(&self)
     }
 
     fn requires_notarization(&self) -> bool {
-        UtxoMmCoinOpsStandardImpl::requires_notarization(&self)
+        utxo_common::requires_notarization(&self)
     }
 
     fn set_required_confirmations(&self, confirmations: u64) {
-        UtxoMmCoinOpsStandardImpl::set_required_confirmations(&self, confirmations)
+        utxo_common::set_required_confirmations(&self, confirmations)
     }
 
     fn set_requires_notarization(&self, requires_nota: bool) {
-        UtxoMmCoinOpsStandardImpl::set_requires_notarization(&self, requires_nota)
+        utxo_common::set_requires_notarization(&self, requires_nota)
     }
 }
